refactor: 🔄 remove rest folders
